[refactor] #20 - rename project
같은 의미의 보다 짧고 간결한 단어로 프로젝트 명칭을 변경한다.
`purchase -> buy`
