+++ src/main/java/com/lottog/buyer/controller/LoginController.java
@@ -1,13 +1,13 @@
-package com.lottog.purchaser.controller;
+package com.lottog.buyer.controller;
 
-import com.lottog.purchaser.dto.request.LoginRequest;
-import com.lottog.purchaser.dto.response.ErrorResponse;
-import com.lottog.purchaser.dto.response.LoginResponse;
-import com.lottog.purchaser.dto.response.UserInfoResponse;
-import com.lottog.purchaser.service.LoginService;
-import com.lottog.purchaser.service.DepositService;
-import com.lottog.purchaser.service.PurchasableCountService;
-import com.lottog.purchaser.service.SeleniumService;
+import com.lottog.buyer.dto.request.LoginRequest;
+import com.lottog.buyer.dto.response.ErrorResponse;
+import com.lottog.buyer.dto.response.LoginResponse;
+import com.lottog.buyer.dto.response.UserInfoResponse;
+import com.lottog.buyer.service.LoginService;
+import com.lottog.buyer.service.DepositService;
+import com.lottog.buyer.service.PurchasableCountService;
+import com.lottog.buyer.service.SeleniumService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
