+++ src/main/java/com/lottog/buyer/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.lottog.purchaser.dto.request;
+package com.lottog.buyer.dto.request;
 
 public record LoginRequest(
         String id,
