+++ src/main/java/com/lottog/buyer/dto/request/PaymentRequest.java
@@ -1,4 +1,4 @@
-package com.lottog.purchaser.dto.request;
+package com.lottog.buyer.dto.request;
 
 public record PaymentRequest(
         String id
