+++ src/main/java/com/lottog/buyer/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.lottog.purchaser.dto.response;
+package com.lottog.buyer.dto.response;
 
 import org.springframework.http.HttpStatus;
