+++ src/main/java/com/lottog/buyer/dto/response/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package com.lottog.purchaser.dto.response;
+package com.lottog.buyer.dto.response;
 
 public record UserInfoResponse(
         Boolean purchasable,
