+++ src/main/java/com/lottog/buyer/service/LoginService.java
@@ -1,7 +1,7 @@
-package com.lottog.purchaser.service;
+package com.lottog.buyer.service;
 
-import com.lottog.purchaser.dto.request.LoginRequest;
-import com.lottog.purchaser.dto.response.LoginResponse;
+import com.lottog.buyer.dto.request.LoginRequest;
+import com.lottog.buyer.dto.response.LoginResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.WebElement;
