+++ src/main/java/com/lottog/buyer/service/PurchasableCountService.java
@@ -1,4 +1,4 @@
-package com.lottog.purchaser.service;
+package com.lottog.buyer.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
